// capabilities/generated/webapp-v1alpha1.ts
// This file is auto-generated by kubernetes-fluent-client, do not edit manually
import { GenericKind, RegisterKind } from "kubernetes-fluent-client";
export class WebApp extends GenericKind {
  /**
   * WebAppSpec defines the desired state of WebApp
   */
  spec?: Spec;
  /**
   * WebAppStatus defines the observed state of WebApp
   */
  status?: Status;
}

/**
 * WebAppSpec defines the desired state of WebApp
 */
export interface Spec {
  /**
   * Langauge of the web application (english or spanish)
   */
  language: string;
  /**
   * Number of replicas for the deployment
   */
  replicas: number;
  /**
   * Color theme of the web application (dark or light)
   */
  theme: string;
}

/**
 * WebAppStatus defines the observed state of WebApp
 */
export interface Status {
  /**
   * Conditions describing the current state
   */
  conditions?: Condition[];
}

/**
 * Condition contains details for one aspect of the current state of this API Resource.
 */
export interface Condition {
  /**
   * lastTransitionTime is the last time the condition transitioned from one status to
   * another. This is not guaranteed to be set in happensBefore order across different
   * conditions for a given object. It may be unset in some circumstances.
   */
  lastTransitionTime: Date;
  /**
   * observedGeneration represents the .metadata.generation that the condition was set based
   * upon. For instance, if .metadata.generation is currently 12, but the
   * .status.conditions[x].observedGeneration is 9, the condition is out of date with respect
   * to the current state of the instance.
   */
  observedGeneration: number;
  /**
   * reason contains a programmatic identifier indicating the reason for the condition's last
   * transition. If this value is empty, it should be treated as "incomplete".
   */
  reason: string;
  /**
   * status of the condition, Pending, Ready, Failed
   */
  status: string;
}

RegisterKind(WebApp, {
  group: "application.pepr.dev",
  version: "v1alpha1",
  kind: "WebApp",
  plural: "webapps",
});